SOLIDs on relay.js
